Split out UIKit color resource
